Amended issue with lambda's filter to make handling more generic.
When using the lambda to handle termination notifications on accounts that have more than one TargetGroup associated, the filtering was overriding the Name/Value. This was causing the describeTags to have entries not relevant for consideration and resulted in the deregister_target call to be invoked with instance-id's that were not part of the Target group.

The impact was that when using more than one Target Group, instance removal from the target group was non-deterministic, meaning that sometimes the right group would be picked up and the state will go into 'draining', sometimes it was killed without being drained.